Install statsmodels, partial refactor with sm
Finding out that there are features/functions that are in R and not mirrored in scipy stats. Also not necessarily in numpy, or scikitlearn. Seems like translating the R code and features needs a few Python libraries. Best case may be using scipy with statsmodels. Scipy has stats built in, but some of the functions are not the same. Statsmodels seeming has R like functions, but missing things like distributions.

Might be possible to stay in numpy, but have doubts.

TO DO:
- Refactor Chapter 2 Lab and Exercises
- Refactor Chapter 3 Lab
- Make executive decision on future libraries